Introduce PLN opaque type for all monetary case class fields
Zero-cost compile-time type safety: 87 files, ~25 source + ~40 test files.
All monetary Double fields across Firm, Household, BankingSector,
GovState, BankState, ForexState, BopState, NbpState, SfcCheck,
World, and all market/sector state types now use PLN opaque type.
Pattern: .toDouble at read boundaries, PLN() at write boundaries.
1250 tests pass, 0 failures.

+++ src/main/scala/sfc/engine/SectoralMobility.scala
@@ -53,7 +53,7 @@ object SectoralMobility:
     for hh <- households do
       hh.status match
         case HhStatus.Employed(_, sectorIdx, wage) =>
-          sums(sectorIdx.toInt) += wage
+          sums(sectorIdx.toInt) += wage.toDouble
           counts(sectorIdx.toInt) += 1
         case _ =>
     val result = new Array[Double](6)
